+++ Magic/CHANGELOG.md
@@ -6,6 +6,7 @@
  - All builtin mobs using zombie pigmen changed to husks (RIP)
  - Add support for class-locked armor
  - Add "interact" mob trigger for more advanced right-click functionality
+ - Prevent putting wands into a grindstone
 
 # 7.9.7